+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - WMT20 Tests
 - Paracrawl_v5_1 for Pashto and Khmer -English
 - NunavutHansard_v3 for Inuktitut -English
+- 'mtdata -b' for short outputs and crash on error input
 
 # 0.2.8 - 20210126
 - Paracrawl v7 and v7.1  -- 29 new datasets